% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{Abstract_PlotAdd}
\alias{Abstract_PlotAdd}
\title{adds a ggplot component}
\usage{
Abstract_PlotAdd(GGObj, GGfunc, YVal, YcolorName = NULL)
}
\arguments{
\item{GGObj}{a ggplot object we are adding to}

\item{GGfunc}{what gg type object used}

\item{YVal}{name of the y variable used}

\item{YcolorName}{name of the color, either a factor or a string}
}
\value{
GGObj with the appended graphic
}
\description{
adds a ggplot component
}
