% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{Aux_info_data}
\alias{Aux_info_data}
\title{Auxiliary data}
\format{
Rows: 999   Columns: 17
}
\usage{
data("Aux_info_data")
}
\description{
auxiliary information that has been split from the WasteWater_data. Can be 
re-merged using merge(WasteWater_data, Aux_info_data, by="sample_id")
}
\keyword{datasets}
