% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--data_prep--calculated_tools--build_analysis_DF.R
\name{buildWasteAnalysisDF}
\alias{buildWasteAnalysisDF}
\title{Convert wastewater_data data to workset4 shape}
\usage{
buildWasteAnalysisDF(df)
}
\arguments{
\item{df}{data frame object from data/wastewater_data.rda}
}
\value{
data frame
}
\description{
This takes the wastewater_data dataframe and rename variables, 
calculates sars_cov2_adj_load_log10 column, 
and filters rows where average_flow_rate is NA
}
\examples{
data(WasteWater_data, package = "Covid19Wastewater")
data(Pop_data, package = "Covid19Wastewater")
buildWasteAnalysisDF(dplyr::left_join(head(WasteWater_data), Pop_data))
}
