% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildWorkSheet4.R
\name{buildWorkSheet4}
\alias{buildWorkSheet4}
\title{Convert wastewater_data data to workset4 shape}
\usage{
buildWorkSheet4(df)
}
\arguments{
\item{df}{data frame object from data/wastewater_data.rda}
}
\value{
data frame
}
\description{
This takes the wastewater_data dataframe and rename variables, 
calculates sars_cov2_adj_load_log10 column, 
and filters rows where average_flow_rate is NA
}
\examples{
data(DHSWaste_data, package = "DSIWastewater")
buildWorkSheet4(DHSWaste_data)
}
