% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{Case_data}
\alias{Case_data}
\title{Case data}
\format{
Rows: 33103  Columns: 4
}
\usage{
data("Case_data")
}
\description{
Data from the DHS of positive covid tests in Wisconsin cites
}
\keyword{datasets}
