% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_function.R
\name{ClassifyCaseRegression}
\alias{ClassifyCaseRegression}
\title{Create Case Flags based on regression slope}
\usage{
ClassifyCaseRegression(DF)
}
\arguments{
\item{DF}{dataframe that contains results of buildRegressionEstimateTable}
}
\value{
DF with an three extra column Category containing the case flags
case_flag: when the 7 day slope is above 5
case_flag_plus_comm.threshold: when case flag and more then 200 cases
slope_switch_flag: the first case flags in consecutive case flags
}
\description{
Create Case Flags based on regression slope
}
