% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_function.R
\name{classifyQuantileFlagRegression}
\alias{classifyQuantileFlagRegression}
\title{Classify FlagRegression  with rolling Quantile info}
\usage{
classifyQuantileFlagRegression(DF, pval = 0.3)
}
\arguments{
\item{DF}{dataframe that contains results of buildRegressionEstimateTable and
makeQuantileColumns}

\item{Pval}{threashold needed for flag_ntile_pval to flag}
}
\value{
DF with three extra columns 
cdc_flag: when the CDC method labels as 'major increase'
flag_ntile: when the cdc flag and its in the top quantile
flag_ntile_pval: when the flag ntile and the regression slope is less
than pval
}
\description{
Create wastewater flags based on the CDC classification defined in 
classifyRegressionAnalysis and the quantile rank of the date.
}
