% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--analysis_tools--DHS_methods--classification_function.R
\name{classifyRegressionAnalysis}
\alias{classifyRegressionAnalysis}
\title{classifyRegressionAnalysis}
\usage{
classifyRegressionAnalysis(
  DF,
  PSigTest = TRUE,
  per_changed_column = modeled_percentchange,
  model_sig_column = lmreg_sig
)
}
\arguments{
\item{DF}{dataframe that contains results of buildRegressionEstimateTable}

\item{PSigTest}{Controls if we filter values with P-values>.3}

\item{per_changed_column}{column name of estimated percent change of signal}

\item{model_sig_column}{column name of significance of linear model that generated 
the percent change}
}
\value{
DF with an extra column Catagory containing the results of the DHS binning
}
\description{
Adds the DHS Classification scheme to data created by runRegressionAnalysis
}
\examples{
data(Example_data, package = "DSIWastewater")
Example_data$modeled_percentchange = 0
Example_data$lmreg_sig = .01
classifyRegressionAnalysis(Example_data)
}
