% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--data_prep--filter_tools--outlier_tools.R
\name{computeJumps}
\alias{computeJumps}
\title{compute first difference Jumps for N1 and N2}
\usage{
computeJumps(df, N1_column = N1, N2_column = N2, site_column = site)
}
\arguments{
\item{df}{DataFrame. needs Column n1_sars_cov2_conc, n2_sars_cov2_conc, site}

\item{N1_column}{N1 metric used in finding difference}

\item{N2_column}{N2 metric used in finding difference}

\item{site_column}{Grouping that makes each group a time seriecomputeJumpss}
}
\value{
dataframe with 4 columns appended: delta(n1), delta(n2) from left and right
}
\description{
compute first difference Jumps for N1 and N2
}
\examples{
data(Example_data, package = "Covid19Wastewater")
Example_data$site = "Madison"
computeJumps(Example_data)
}
