% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_flag_count.R
\name{Count_Flag}
\alias{Count_Flag}
\title{Create counts of flag data}
\usage{
Count_Flag(DF, group = c("Site", "window", "quant"))
}
\arguments{
\item{DF}{data frame with grouping columns and numeric flag columns}

\item{group}{vector specifying what columns should be used for grouping}
}
\value{
DF with the number of flags of each type with in groups
}
\description{
Takes a data frame with grouping columns and numeric flag DF where 1 means 
a flag
}
