% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--analysis_tools--DHS_methods--create_flag_count.R
\name{countFlags}
\alias{countFlags}
\title{Create counts of flag data}
\usage{
countFlags(DF, group = c("Site", "window", "quant"))
}
\arguments{
\item{DF}{data frame with grouping columns and numeric flag columns}

\item{group}{vector specifying what columns should be used for grouping}
}
\value{
DF with the number of flags of each type with in groups
}
\description{
Takes a data frame with grouping columns and numeric flag DF where 1 means 
a flag
}
\examples{
data(Example_data, package = "Covid19Wastewater")
countFlags(Example_data, group = c("site"))
}
