% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{Covariants_data}
\alias{Covariants_data}
\title{Covariants data}
\format{
Rows: 69   Columns: 33
}
\usage{
data("Covariants_data")
}
\description{
dataframe containing info about the proportion of each COVID-19 variant
every 2 weeks. This data is from 
[https://github.com/hodcroftlab/covariants/tree/master](https://github.com/hodcroftlab/covariants/tree/master)
}
\keyword{datasets}
