% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19WasteWater.R
\docType{package}
\name{Covid19Wastewater}
\alias{Covid19Wastewater}
\title{Covid19Wastewater: A package for running Covid19 wastewater concentration analysis}
\description{
This is an R package of utilities to perform wastewater data analysis for pathogenic surveillance and monitoring.
}
\details{
This project is a collaboration between the University of Wiscosin's \href{https://datascience.wisc.edu}{Data Science Institute (DSI)}, the \href{https://www.dhs.wisconsin.gov/covid-19/wastewater.htm}{Wisconsin Department of Health Services (DHS)}, and the \href{https://www.slh.wisc.edu/environmental/covid-19-wastewater/}{State Lab of Hygiene (SLH)}.

If you need help getting started view our \href{getting started guide}{https://github.com/UW-Madison-DSI/Covid19Wastewater/blob/main/docs/getting-started/Starting_Guide.md}
}
\section{Built in data}{

View our data vignette to learn more about our built in data.
To view what each column name means view our \href{Data_Description}{https://github.com/UW-Madison-DSI/Covid19Wastewater/blob/main/docs/data/data_columns_discription.md}
vignette(HFG_data_case)
vignette(longitudinal_data_case)
vignette(variant_data)
vignette(HFG_data_waste)
vignette(aux_info)
vignette(longitudinal_data_waste)
vignette(Intercepter_data_case)
vignette(example_data)
vignette(population_data)
}

\section{Data Prep}{

vignette(calculated_info)
vignette(outliers)
vignette(smoothing)
}

\section{Analysis tools}{

vignette(time_series_offset)
vignette(linear_forest)


View more in our github repo: \href{https://github.com/UW-Madison-DSI/Covid19Wastewater}{Here}
}

