% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{createFlagGraph_plot}
\alias{createFlagGraph_plot}
\title{Create graphics showing flags as vertical lines}
\usage{
createFlagGraph_plot(
  MainDF,
  FlagDF,
  Flag1 = NULL,
  Flag2 = NULL,
  xVal = "date",
  PointVal = NULL,
  LineVal = NULL,
  facetFormula = " ~ Site"
)
}
\arguments{
\item{MainDF}{}

\item{FlagDF}{}

\item{Flag1}{}

\item{Flag2}{}

\item{xVal}{}

\item{PointVal}{}

\item{LineVal}{}

\item{facetFormula}{}
}
\description{
Create graphics showing flags as vertical lines
}
