% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{CreateHeatMaps_Plot}
\alias{CreateHeatMaps_Plot}
\title{CreateHeatMaps_Plot}
\usage{
CreateHeatMaps_Plot(DF, FacGridFormula, FillFac, CatagoryColors)
}
\arguments{
\item{DF}{The DF used to create the Heatmap}

\item{FacGridFormula}{how we wan to facet the heatmap}

\item{FillFac}{the name of the category method}

\item{CatagoryColors}{The color scheme used}
}
\value{
faceted ggplot heatmap
}
\description{
Creates graphic of model prediction for each method
}
\keyword{internal}
