% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library--analysis_tools--plot--plot_functions.R
\name{createMethodCompareBar_Plot}
\alias{createMethodCompareBar_Plot}
\title{createMethodCompareBar_Plot: Compare Regression analysis methods}
\usage{
createMethodCompareBar_Plot(DF, Method = "Method", Cat = "Catagory")
}
\arguments{
\item{DF}{data frame containing all the analysis for every method}

\item{Method}{factor column for each type of regression done}

\item{Cat}{Category column that the regression analysis is store}
}
\value{
ggplot object
}
\description{
createMethodCompareBar_Plot: Compare Regression analysis methods
}
\examples{
data("example_data", package = "DSIWastewater")
example_data$Method <- sample(c("MethodA", "MethodB"), 7, replace = TRUE)
example_data$Catagory <- sample(c("major decrease", "major increase"), 7, replace = TRUE)
DSIWastewater:::createMethodCompareBar_Plot(example_data)
}
\keyword{internal}
