% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_functions.R
\name{CreateWasteFlags}
\alias{CreateWasteFlags}
\title{Title}
\usage{
CreateWasteFlags(DF, windows = c(14, 30, 60, 90), quants = c(5:9)/10)
}
\arguments{
\item{DF}{}

\item{windows}{}

\item{quants}{}
}
\description{
Title
}
