% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src--analysis_tools--plot--plot_functions.R
\name{createWasteGraph_Plot}
\alias{createWasteGraph_Plot}
\title{Wastewater graphic}
\usage{
createWasteGraph_Plot(
  DF,
  xVal,
  PointVal = NULL,
  LineVal = NULL,
  facetFormula = "Data ~ site"
)
}
\arguments{
\item{DF}{DF containing wastewater measurements specified in the remaining params}

\item{xVal}{name of x variable, normally close to "Date"}

\item{PointVal}{the discrete measurements}

\item{LineVal}{the continuous measurements}

\item{facetFormula}{formula of how to facet the plot}
}
\value{
a ggplot object with points with lables for each PointVal and a lines for each LineVal
}
\description{
Wastewater graphic
}
\examples{
data("example_data", package = "DSIWastewater")
DSIWastewater:::createWasteGraph_Plot(example_data, "date", LineVal = "geoMean",
                                     facetFormula  = "~WWTP")
}
\keyword{internal}
