% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Description.R
\name{Data_Description}
\alias{Data_Description}
\title{Data Description
This package contains a lot of data with many column names, here is a list of them all: [https://github.com/UW-Madison-DSI/Covid19Wastewater/blob/main/docs/data/data_columns_discription.md](https://github.com/UW-Madison-DSI/Covid19Wastewater/blob/main/docs/data/data_columns_discription.md)}
\description{
Data Description
This package contains a lot of data with many column names, here is a list of them all: [https://github.com/UW-Madison-DSI/Covid19Wastewater/blob/main/docs/data/data_columns_discription.md](https://github.com/UW-Madison-DSI/Covid19Wastewater/blob/main/docs/data/data_columns_discription.md)
}
