% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--data_prep--calculated_tools--date_vector_difference.R
\name{date_distance_calc}
\alias{date_distance_calc}
\title{date_distance_calc}
\usage{
date_distance_calc(DF, base_date_vec, vecNames)
}
\arguments{
\item{DF}{DF to extract vector from}

\item{base_date_vec}{date flag vector to pull each distance from}

\item{vecNames}{a vector of column vec names each with binary flagging info}
}
\value{
DF containing the distance of each term to the base vector
}
\description{
date_distance_calc
}
\examples{
data("Example_data", package = "Covid19Wastewater")
Example_data$Late_date <- sample(Example_data$date)
Example_data$Late_date[sample(1:length(Example_data), length(Example_data) / 3)] <- NA
head(date_distance_calc(Example_data, "date", "Late_date"))
}
