% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--data_prep--calculated_tools--date_vector_difference.R
\name{DF_date_vector}
\alias{DF_date_vector}
\title{DF_date_vector}
\usage{
DF_date_vector(DF, date_vec, flag_vecs)
}
\arguments{
\item{DF}{DF containing the other vectors as columns}

\item{date_vec}{column vec name with date dataType}

\item{flag_vecs}{a vector of column vec names each with binary flagging info}
}
\value{
an equivilent DF where the 1 in the flag info is replaced with the date
}
\description{
DF_date_vector
}
\examples{
data("Example_data", package = "DSIWastewater")
Example_data$flag = 1
DF_date_vector(Example_data, date, "flag")
}
