% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--data_prep--calculated_tools--date_vector_difference.R
\name{diffLookup}
\alias{diffLookup}
\title{lookup}
\usage{
diffLookup(DFCol, base_date_vec)
}
\arguments{
\item{DFCol}{date flag vector used to find min distance}

\item{base_date_vec}{date flag vector used to find baseline min distance}
}
\value{
difference between DFCol and the closest entry of base_date_vec
}
\description{
lookup
}
\examples{
data("Example_data", package = "DSIWastewater")
Example_data$Late_date <- sample(Example_data$date)
Example_data$Late_date[sample(1:length(Example_data), length(Example_data) / 3)] <- NA
head(DSIWastewater:::diffLookup(Example_data$date, Example_data$Late_date))
}
\keyword{internal}
