% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSIWasteWater.R
\docType{package}
\name{DSIWastewater}
\alias{DSIWastewater}
\title{DSIWastewater: A package for running Covid wastewater concentration analysis}
\description{
The DSIWastewater package provides three categories of important functions:
DataTableGen, DataMod and Viz
}
\section{DataTableGen functions}{

This section list the functions used for running the DHS analysis

buildRegressionEstimateTable - This function is the composition of the next two

runRegressionAnalysis

ClassifyRegressionAnalysis
}

\section{DataMod functions}{

buildWasteAnalysisDF
}

\section{Viz functions}{

createRegressionAnalysis_Plot
}

