% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Toy example of worksheet 4 to be used in package examples}
\format{
Rows: 7 Columns: 3
}
\usage{
data("example_data")
}
\description{
Toy example of worksheet 4 to be used in package examples
}
\keyword{datasets}
