% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/src--analysis_tools--lm_rand_forest--formula_help.R
\name{expand_formula}
\alias{expand_formula}
\title{Expand formula for increased info
takes a formula with shape A ~ B | C 
and convert . to its real representation}
\usage{
expand_formula(X, data)
}
\arguments{
\item{X}{formula object like A ~ B | C}

\item{data}{data object to extract . info from}
}
\value{
formula object
}
\description{
Expand formula for increased info
takes a formula with shape A ~ B | C 
and convert . to its real representation
}
