% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aug--ts_trend_gen.R
\name{expSmoothMod}
\alias{expSmoothMod}
\title{expSmoothMod
Add a column of the smoothed values using exponential smoothing}
\usage{
expSmoothMod(DF, InVar, OutVar, alpha = "guess", beta = "guess", Filter = NULL)
}
\arguments{
\item{DF}{The DF we are to add a exponential smoothing column to}

\item{InVar}{The column to be smoothed}

\item{OutVar}{The name of the new column}

\item{alpha}{The alpha fed into robets exponential smoothing. if it equals "guess" then it
if found using parameterGuess}

\item{beta}{The beta fed into robets exponential smoothing. if it equals "guess" then it
if found using parameterGuess}

\item{Filter}{Prefilter using the value of a Filter col}
}
\value{
A DF with an extra col with a exp smoothed version of InVar
}
\description{
expSmoothMod
Add a column of the smoothed values using exponential smoothing
}
\keyword{"DSIWastewater")}
\keyword{'Algoma}
\keyword{<-}
\keyword{=}
\keyword{==}
\keyword{DSIWastewater:::expSmoothMod(examp_DF,"n1_sars_cov2_conc","expN1"'}
\keyword{WWTF",]}
\keyword{buildWasteAnalysisDF(WasteWater_data)}
\keyword{data("WasteWater_data",}
\keyword{examp_DF}
\keyword{examp_DF[examp_DF$site}
\keyword{internal}
\keyword{package}
