% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_vec_by_number_points.R
\name{FactorVecByNumPoints}
\alias{FactorVecByNumPoints}
\title{convert column to factor based on amount of entries}
\usage{
FactorVecByNumPoints(DF, FacVar, FiltVar = NA)
}
\arguments{
\item{DF}{The data frame being manipulated}

\item{FacVar}{The column being converted to a factor}

\item{FiltVar}{A column to filter before counting entrys.
Defaults to FacVar}
}
\value{
DF with the FacVar column a factor
}
\description{
convert column to factor based on amount of entries
}
