% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_vec_by_number_points.R
\name{FactorVecByVec}
\alias{FactorVecByVec}
\title{Title}
\usage{
FactorVecByVec(FactorDF, OrderDF, FacVar, NumVar)
}
\arguments{
\item{FactorDF}{}

\item{OrderDF}{}

\item{FacVar}{}

\item{NumVar}{}
}
\description{
Title
}
