% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_rand_forest--info.R
\name{gen_INCMSE}
\alias{gen_INCMSE}
\title{get increased mean square error for each column}
\usage{
gen_INCMSE(tree_model)
}
\arguments{
\item{tree_model}{random_linear_forest object you want the MSE of}
}
\value{
data.frame containing each column and its MSE increase
}
\description{
get increased mean square error for each column
}
\examples{
data("example_data", package = "DSIWastewater")
model <- random_linear_forest(example_data, 2, PMMoV ~ N1 + N2 | date + site)
gen_OOB_pred(model)
}
