% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library--analysis_tools--lm_rand_forest--info.R
\name{gen_OOB_pred}
\alias{gen_OOB_pred}
\title{get OOB predictions of the training dataset
returns the predictions of each row of the input data
using only trees not trained on the row}
\usage{
gen_OOB_pred(tree_model, incMSE = NA, resid = FALSE)
}
\arguments{
\item{tree_model}{random_linear_forest object you want the OOB predictions of}

\item{incMSE}{if its a numeric scramble the given column}

\item{resid}{if True return the residuals of the model instead}
}
\value{
data.frame or vector of oob predictions
}
\description{
get OOB predictions of the training dataset
returns the predictions of each row of the input data
using only trees not trained on the row
}
\examples{
data(Example_data, package = "DSIWastewater")
model <- random_linear_forest(Example_data, 2, PMMoV ~ N1 + N2 | pcr_type)
gen_OOB_pred(model, resid = TRUE)
}
