% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--analysis_tools--time_offset--Offset_Analysis.R
\name{heatmapcorfunc}
\alias{heatmapcorfunc}
\title{Given output from OffsetDFMaker returns a 2x3 grid of all the plots with highlighted values}
\usage{
heatmapcorfunc(cordata)
}
\arguments{
\item{cordata}{DF with geo_mean and conf_case columns}

\item{data}{Output from OffsetDFMaker}

\item{title}{Title you want (appears at bottom)}
}
\value{
ggplot plot object

ggplot plot object (heatmap)
}
\description{
Given output from OffsetDFMaker returns a 2x3 grid of all the plots with highlighted values
}
\examples{
data(WasteWater_data, package = "DSIWastewater")
data("Case_data", package = "DSIWastewater")
OffsetDFMaker_Output <- OffsetDFMaker(10,as.Date("2020-08-01"), as.Date("2023-01-01"), Case_data, WasteWater_data)
OffsetDF_Plot(OffsetDFMaker_Output,"All Wisconsin data over all time")
Outputs a heatmap where the color is the r squared of wastewater data and center day + x many future days and y many past days
Helps inform Offset Analysis
}
