% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{HFGCase_data}
\alias{HFGCase_data}
\title{Case data set from the SLH on 2/7/22}
\format{
Rows: 788  Columns: 6
}
\usage{
data("HFGCase_data")
}
\description{
Data from the DHS of positive covid tests in selected Wisconsin cites where
high frequency testing of covid data was done
}
\details{
Raw file 'HighFreq_CaseData_2021-05-07.csv'
From Steve
}
\keyword{datasets}
