% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{HFGWaste_data}
\alias{HFGWaste_data}
\title{High frequency Waste data}
\format{
Rows: 3078   Columns: 18
}
\usage{
data("HFGWaste_data")
}
\description{
Data from the DHS of  HFG covid wastewater concentration
}
\details{
Raw file 'HFG_data_corrected_w_HF183&CrP_011422.xlsx'
Dagmara email Fri 2/25/2022 6:15 PM
}
\keyword{datasets}
