% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{InterceptorCase_data}
\alias{InterceptorCase_data}
\title{Madison interceptor Case data set}
\format{
Rows: 3288   Columns: 5
}
\usage{
data("InterceptorCase_data")
}
\description{
Data from the DHS of positive covid tests for Madison interceptors
}
\details{
Raw file 'MSN_Interceptor_Cases.csv'
From email Nathan on 2/4/2022
}
\keyword{datasets}
