% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeQuantileFunction.R
\name{MakeQuantileColumns}
\alias{MakeQuantileColumns}
\title{Title}
\usage{
MakeQuantileColumns(DF, column, quants, windows)
}
\arguments{
\item{DF}{}

\item{column}{}

\item{quants}{}

\item{windows}{}
}
\description{
Title
}
