% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--data_prep--smooth_tools--ts_trend_gen.R
\name{nGuess}
\alias{nGuess}
\title{nGuess for sgolaySmoothMod number of points per polynomial}
\usage{
nGuess(DF, InVar, Base, min)
}
\arguments{
\item{DF}{The DF to fit the sgolayfilt curve on. should only
be from one TS}

\item{InVar}{The col name that the variable is about to smooth}

\item{Base}{The value we are scaling by the of the number of
rows in the data}

\item{min}{The smallest value returned}
}
\value{
a number greater or equal to min
}
\description{
nGuess for sgolaySmoothMod number of points per polynomial
}
\examples{
data(example_data, package = "DSIWastewater")
DSIWastewater:::nGuess(example_data, "geoMean", 50/178, 7)
}
\keyword{internal}
