% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot--plot_functions.R
\name{orderAndZipListsOfPlots_Plot}
\alias{orderAndZipListsOfPlots_Plot}
\title{Take two list of plots and combine them into one 3 col long plot}
\usage{
orderAndZipListsOfPlots_Plot(
  top_plot_list,
  bot_plot_list,
  ratA = 3,
  ratB = 1,
  nbreak = 3,
  IsLong = TRUE
)
}
\arguments{
\item{top_plot_list}{Lists of plots that get added on top, generically the 
rectangle plot of method prediction. Needs to be the same length as bot_plot_list}

\item{bot_plot_list}{List of plots to be combined on the bottom}

\item{ratA}{The proportion the top plot should be.}

\item{ratB}{The proportion the bot plot should be.}

\item{nbreak}{Where the plot should be faceted}

\item{IsLong}{Controls whether the plot is wide or long}
}
\value{
a ggplot
}
\description{
Take two list of plots and combine them into one 3 col long plot
}
\keyword{Used}
\keyword{createRegressionAnalysis_Plot.}
\keyword{for}
\keyword{in}
\keyword{internal}
\keyword{normal}
\keyword{not}
\keyword{recommended}
\keyword{use.}
