% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--data_prep--smooth_tools--ts_trend_gen.R
\name{parameterGuess}
\alias{parameterGuess}
\title{Get a fitting parameter for the model}
\usage{
parameterGuess(DF, InVar, Base, max)
}
\arguments{
\item{DF}{The DF we are about to fit a smoothing curve on. should only
be from one TS}

\item{InVar}{The col name that the variable is about to smooth}

\item{Base}{The value we are scaling by the inverse of the number of
rows in the data}

\item{max}{The largest value returned}
}
\value{
a number between 0 and max
}
\description{
Get a fitting parameter for the model
}
\examples{
data(Example_data, package = "Covid19Wastewater")
Covid19Wastewater::parameterGuess(Example_data,"geo_mean", 17.8, .6)
}
\keyword{internal}
