% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library--analysis_tools--lm_rand_forest--use.R
\name{predict,random_linear_forest-method}
\alias{predict,random_linear_forest-method}
\title{predict new data from random_linear_forest models}
\usage{
\S4method{predict}{random_linear_forest}(object, new_data)
}
\arguments{
\item{random_linear_forest}{}
}
\value{
vector of predictions for each row
}
\description{
predict new data from random_linear_forest models
}
\examples{
data("example_data", package = "DSIWastewater")
model <- random_linear_forest(PMMoV ~ N1 + N2 | date + site)
predict(model, example_data)
}
