% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_tools--lm_rand_forest--class_methods.R
\docType{class}
\name{random_linear_forest-class}
\alias{random_linear_forest-class}
\title{random_linear_forest
model class using a random forest of linear forest models}
\value{
random_linear_forest object
}
\description{
random_linear_forest
model class using a random forest of linear forest models
}
\section{Slots}{

\describe{
\item{\code{formula}}{formula used to create trees and linear models}

\item{\code{models}}{list of each tree model in model}

\item{\code{data}}{data.frame of the input data}

\item{\code{resid}}{numeric vector of residuals}

\item{\code{inbag_data}}{list of data.frame each model trained on}

\item{\code{oob_data}}{list of data.frame containing info model not trained on}

\item{\code{oob_resid}}{numeric vector of oob residuals}
}}

\examples{
data("example_data", package = "DSIWastewater")
random_linear_forest(example_data, 2, PMMoV ~ N1 + N2 | date + site)
}
