% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--analysis_tools--DHS_methods--regression_algo.R
\name{runRegressionAnalysis}
\alias{runRegressionAnalysis}
\title{runRegressionAnalysis}
\usage{
runRegressionAnalysis(DF, Formula, Keep = NULL, n = 5, verbose = FALSE)
}
\arguments{
\item{DF}{Contains TS data should only contain one site}

\item{Formula}{LM model to be fit on a subsection of the data}

\item{Keep}{The col in the original DF to keep besides Date}

\item{n}{number of rows to be in each LM regression}

\item{verbose}{prints what site we are in}
}
\value{
a row of a DF containing the 
site, last date, timespan, number of rows, model slope and significance,
and predicted percent change, and what linear model was used
}
\description{
The DHS model system. runs a LM on each each set of n consecutive measurements
and returns a summary of the information
}
\examples{
data(Example_data, package = "DSIWastewater")
DSIWastewater:::runRegressionAnalysis(Example_data, Formula = geoMean ~ date)
}
\keyword{internal}
