% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library--analysis_tools--lm_rand_forest--use.R
\name{summary,random_linear_forest-method}
\alias{summary,random_linear_forest-method}
\title{summary method for linear forest class}
\usage{
\S4method{summary}{random_linear_forest}(object)
}
\arguments{
\item{random_linear_forest}{}
}
\value{
summary.random_linear_forest object
}
\description{
summary method for linear forest class
}
\examples{
data("example_data", package = "DSIWastewater")
model <- random_linear_forest(PMMoV ~ N1 + N2 | date + site)
summary(model)
}
