% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--analysis_tools--DHS_methods--regression_algo.R
\name{uniqueVal}
\alias{uniqueVal}
\title{Find all unique values in the column selected}
\usage{
uniqueVal(Col, DF)
}
\arguments{
\item{Col}{Col being looked at}

\item{DF}{the DF containing Col}
}
\value{
a list of each unique col value
}
\description{
Find all unique values in the column selected
}
\examples{
data(Example_data, package = "Covid19Wastewater")
Covid19Wastewater:::uniqueVal("site", Example_data)
}
\keyword{internal}
