% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--analysis_tools--time_offset--Offset_Analysis_Plot.R
\name{VariantPlot}
\alias{VariantPlot}
\title{Shows each variant in proportion to the others in 2 week time periods}
\usage{
VariantPlot(covar)
}
\arguments{
\item{covar}{Covariant data frame}
}
\value{
ggplotly object
}
\description{
Shows each variant in proportion to the others in 2 week time periods
}
\examples{
data(Covariants_data, package = "Covid19Wastewater")
VariantPlot(Covariants_data)
}
