% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{WasteWater_data}
\alias{WasteWater_data}
\title{Wastewater data set cleaned from the surveillance branch}
\format{
Rows: 8386 Columns: 30
}
\usage{
data("WasteWater_data")
}
\description{
Data from the DHS of wastewater prevalence in Wisconsin cites
}
\details{
Raw file 'Shareable_Data_wastewater.csv'
From zip file attached to 4/21/2022 email from Matthew Schinwald
}
\keyword{datasets}
