% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--data_prep--filter_tools--make_quantile_function.R
\name{windowingQuantFunc}
\alias{windowingQuantFunc}
\title{create rolling quantile column based on date}
\usage{
windowingQuantFunc(DF, column)
}
\arguments{
\item{DF}{dataframe containing columns:
date: date variable communicating the day the measurement is from
window: singular type column saying what window the rolling function
 should use
quant: singular value column saying what quantile to return
*column: DF needs to contain a column with the same name as the string
in the variable column}

\item{column}{what column to use for the rolling quantile and mean functions}
}
\value{
DF with added columns
ntile: a rolling quantile of the data
pastKavg.wwlog10: a mean of the last K days
}
\description{
Takes 1 time series dataframe and adds a some columns specifying the 
rolling quantile and rolling mean of the selected column.
}
\examples{
data("Example_data", package = "Covid19Wastewater")
Example_data$window = 7
Example_data$quant = .8
head(Covid19Wastewater::windowingQuantFunc(Example_data, "geo_mean"))
}
\keyword{internal}
