% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src--analysis_tools--plot--plot_functions.R
\name{abstract_PlotAdd}
\alias{abstract_PlotAdd}
\title{adds a ggplot component}
\usage{
abstract_PlotAdd(GGObj, GGfunc, YVal, YcolorName = NULL)
}
\arguments{
\item{GGObj}{a ggplot object we are adding to}

\item{GGfunc}{what gg type object used}

\item{YVal}{name of the y variable used}

\item{YcolorName}{name of the color, either a factor or a string}
}
\value{
GGObj with the appended graphic
}
\description{
adds a ggplot component
}
\examples{
data("example_data", package = "DSIWastewater")
library(ggplot2)
GGObj <- ggplot(example_data, aes(x=date))
DSIWastewater:::abstract_PlotAdd(GGObj, geom_line, "geoMean")
}
\keyword{internal}
