% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_analysis_DF.R
\name{buildCaseAnalysisDF}
\alias{buildCaseAnalysisDF}
\title{Prep case data into right format}
\usage{
buildCaseAnalysisDF(df)
}
\arguments{
\item{df}{case dataframe have columns: Date, population_served, FirstConfirmed}
}
\value{
DF with a 7 day rolling sum and a population weighted case column
}
\description{
Prep case data into right format
}
\examples{
data(Case_data, package = "DSIWastewater")
buildCaseAnalysisDF(Case_data)
}
