% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassificationFunction.R
\name{ClassifyCaseRegression}
\alias{ClassifyCaseRegression}
\title{Create Case Flags based on regression slope}
\usage{
ClassifyCaseRegression(DF)
}
\arguments{
\item{DF}{dataframe that contains results of buildRegressionEstimateTable}
}
\value{
DF with an three extra column Catagory containing the case flags
}
\description{
Create Case Flags based on regression slope
}
