% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionAlgo.R
\name{ClassifyRegressionAnalysis}
\alias{ClassifyRegressionAnalysis}
\title{ClassifyRegressionAnalysis}
\usage{
ClassifyRegressionAnalysis(DF, PSigTest = TRUE)
}
\arguments{
\item{DF}{DF that contains results of DHS analysis}

\item{PSigTest}{Controls if we filter values with P-values>.3}
}
\value{
DF with an extra column Catagory containing the results of the DHS binning
}
\description{
Adds the DHS Classification scheme to data created by runRegressionAnalysis
}
