% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_function.R
\name{classifyRegressionAnalysis}
\alias{classifyRegressionAnalysis}
\title{classifyRegressionAnalysis}
\usage{
classifyRegressionAnalysis(DF, PSigTest = TRUE)
}
\arguments{
\item{DF}{dataframe that contains results of buildRegressionEstimateTable}

\item{PSigTest}{Controls if we filter values with P-values>.3}
}
\value{
DF with an extra column Catagory containing the results of the DHS binning
}
\description{
Adds the DHS Classification scheme to data created by runRegressionAnalysis
}
