% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_tools.R
\name{computeJumps}
\alias{computeJumps}
\title{compute first difference Jumps for N1 and N2}
\usage{
computeJumps(df)
}
\arguments{
\item{df}{DataFrame. needs Column n1_sars_cov2_conc, n2_sars_cov2_conc, WWTP}
}
\value{
dataframe with 4 columns appended: delta(n1), delta(n2) from left and right
}
\description{
compute first difference Jumps for N1 and N2
}
\examples{
data(example_data, package = "DSIWastewater")
computeJumps(example_data)
}
