% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19WasteWater.R
\docType{package}
\name{Covid19Wastewater}
\alias{Covid19Wastewater}
\title{Covid19Wastewater: A package for running Covid wastewater concentration analysis}
\description{
This is an R package of utilities to perform wastewater data analysis for pathogenic surveillance and monitoring.
}
\details{
This project is a collaboration between the University of Wiscosin's \href{https://datascience.wisc.edu}{Data Science Institute (DSI)}, the \href{https://www.dhs.wisconsin.gov/covid-19/wastewater.htm}{Wisconsin Department of Health Services (DHS)}, and the \href{http://www.slh.wisc.edu/environmental/covid-19-wastewater}{State Lab of Hygiene (SLH)}.

View more in our github repo: \href{https://github.com/UW-Madison-DSI/Covid19Wastewater}{Here}
}
\section{Built in data}{

\link{Data_Description}
}

\section{Data Prep}{

temp
}

\section{Analysis tools}{

temp
}

