% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{createFlagGraph_plot}
\alias{createFlagGraph_plot}
\title{Create graphics showing flags as vertical lines}
\usage{
createFlagGraph_plot(
  MainDF,
  FlagDF,
  Flag1 = NULL,
  Flag2 = NULL,
  xVal = "date",
  PointVal = NULL,
  LineVal = NULL,
  facetFormula = " ~ site"
)
}
\arguments{
\item{MainDF}{DF used for the non geom_vline calls}

\item{FlagDF}{DF used for ploting the flag vertical lines}

\item{Flag1}{name of first flag}

\item{Flag2}{name of second flag, if only one flag leave NULL}

\item{xVal}{column to use as x var}

\item{PointVal}{list of columns to be shown as points}

\item{LineVal}{list of columns to be shown as lines}

\item{facetFormula}{formula object to be fed into geom_facet}
}
\value{
ggplot object of scatter, line and vertical lines
}
\description{
Create graphics showing flags as vertical lines
}
\examples{
data("example_data", package = "DSIWastewater")
plotDF <- data.frame(
    date = seq(as.Date("2020/9/15"), as.Date("2020/10/6"), 1),
    ExampleFlag = sample(c(1,0,0,0, 0), 22, replace = TRUE),
    site = "Appleton WWTF"
    )
    createFlagGraph_plot(example_data, plotDF, "ExampleFlag", 
                         LineVal = "geoMean")
}
