% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{createRegressionAnalysis_Plot}
\alias{createRegressionAnalysis_Plot}
\title{Create representative plot of the DHS analysis}
\usage{
createRegressionAnalysis_Plot(
  RegDF,
  BaseDF,
  FacGridFormula = Method ~ site,
  PointVal = "sars_cov2_adj_load_log10",
  LineVal = NULL,
  nbreak = 3,
  IsLong = TRUE
)
}
\arguments{
\item{RegDF}{data frame containing the regression analysis}

\item{BaseDF}{The data frame containing the raw data}

\item{FacGridFormula}{The formula we wish to facet the heat maps with}

\item{PointVal}{The point columns we want to plot}

\item{LineVal}{The Line columns we want to plot}

\item{nbreak}{The number of plots in each row}

\item{IsLong}{Controls whether the plot is wide or long}
}
\value{
a ggplot of the heat map of each method and the underlying data
}
\description{
createRegressionAnalysis_Plot uses RegDF to create the top set of plots showing the 
predictions at each time period. 
It uses BaseDF to create the lower plot that shows what the regressed data
looks like
}
\examples{
plotDF <- data.frame(
date = seq(as.Date("2000/10/17"), as.Date("2001/10/17"), 1),
days_elapsed = 1,
Method = sample(c("RandomMethod1", "RandomMethod2"), 366, replace = TRUE),
site = "Appleton WWTF",
Catagory = sample(c("major decrease","no change", "major increase"), 366, replace = TRUE))
names(example_data)[names(example_data) == 'WWTP'] <- 'site'
example_data$Method <- sample(c("RandomMethod1", "RandomMethod2"), 7, replace = TRUE)
createRegressionAnalysis_Plot(plotDF, example_data)
}
