% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aug--date_vector_difference.R
\name{date_distance_clamp}
\alias{date_distance_clamp}
\title{remove distances above threshold}
\usage{
date_distance_clamp(DF, vecNames, thresh)
}
\arguments{
\item{DF}{source dataframe}

\item{vecNames}{column names in DF to be modified}

\item{thresh}{max distance not reduced to thresh from data}
}
\value{
DF with distances above threshold clamped to threshold
}
\description{
remove distances above threshold
}
\examples{
data("example_data", package = "DSIWastewater")
df <- date_distance_calc(example_data, "geoMean", "n")
date_distance_remove(df, "geoMean", 21)
}
