% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_vector_difference.R
\name{diffLookup}
\alias{diffLookup}
\title{lookup}
\usage{
diffLookup(DFCol, base_date_vec, edge = NA)
}
\arguments{
\item{DFCol}{date flag vector used to find min distance}

\item{base_date_vec}{date flag vector used to find baseline min distance}
}
\value{
diffrence between DFCol and the closest entry of base_date_vec
}
\description{
lookup
}
\keyword{internal}
