% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_vector_difference.R
\name{diffLookup}
\alias{diffLookup}
\title{lookup}
\usage{
diffLookup(DFCol, base_date_vec, edge = NA, Method = "Remove")
}
\arguments{
\item{DFCol}{date flag vector used to find min distance}

\item{base_date_vec}{date flag vector used to find baseline min distance}

\item{edge}{threshold to prevent extreme differences from effect results}

\item{Method}{How to deal with var outside of the edge. Clamp if we want to 
make them the edge and Remove if we want to remove them.}
}
\value{
difference between DFCol and the closest entry of base_date_vec
}
\description{
lookup
}
\examples{
data("example_data", package = "DSIWastewater")
DSIWastewater:::diffLookup(example_data$geoMean, example_data$n)
}
\keyword{internal}
