% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{Example_data}
\alias{Example_data}
\title{Example data}
\format{
Rows: 500 Columns: 69
}
\usage{
data("Example_data")
}
\description{
Toy example of full data contains 3 sites and roughly 500 entries 
(waste and case merged together)
}
\keyword{datasets}
