% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--data_prep--calculated_tools--factor_vec_by_number_points.R
\name{factorVecByNumPoints}
\alias{factorVecByNumPoints}
\title{convert column to factor based on amount of entries}
\usage{
factorVecByNumPoints(DF, FacVar, FiltVar = NA)
}
\arguments{
\item{DF}{The data frame being manipulated}

\item{FacVar}{The column being converted to a factor}

\item{FiltVar}{A column to filter before counting entrys.
Defaults to FacVar}
}
\value{
DF with the FacVar column a factor
}
\description{
convert column to factor based on amount of entries
}
\examples{
data("Example_data", package = "Covid19Wastewater")
head(Covid19Wastewater::factorVecByNumPoints(Example_data, "site"))
}
\keyword{internal}
