% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--data_prep--calculated_tools--factor_vec_by_number_points.R
\name{factorVecByVec}
\alias{factorVecByVec}
\title{Get ordering for ploting based on factoring vector}
\usage{
factorVecByVec(FactorDF, OrderDF, FacVar, NumVar)
}
\arguments{
\item{FactorDF}{DF containing FacVar}

\item{OrderDF}{DF containing NumVar}

\item{FacVar}{Column given ordering}

\item{NumVar}{Column used to find ordering}
}
\value{
FactorDF with FacVar being a factor ordered by NumVar
}
\description{
Get ordering for ploting based on factoring vector
}
\examples{
data(Example_data, package = "DSIWastewater")
head(factorVecByVec(Example_data, Example_data, "site", "N1"))
}
