% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/src--data_prep--calculated_tools--factor_vec_by_number_points.R
\name{factorVecByVec}
\alias{factorVecByVec}
\title{Get ordering for ploting based on factoring vector}
\usage{
factorVecByVec(FactorDF, OrderDF, FacVar, NumVar)
}
\arguments{
\item{FactorDF}{DF containing FacVar}

\item{OrderDF}{DF containing NumVar}

\item{FacVar}{Column given ordering}

\item{NumVar}{Column used to find ordering}
}
\value{
FactorDF with FacVar being a factor ordered by NumVar
}
\description{
Get ordering for ploting based on factoring vector
}
\examples{
data("example_data", package = "DSIWastewater")
factorVecByVec(example_data, example_data, "WWTP", "n1_sars_cov2_conc")
}
