% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/src--data_prep--filter_tools--sd_trend_outlier.R
\name{Flag_From_Trend}
\alias{Flag_From_Trend}
\title{Flag values as outliers based on error from estimated trend
This function can be done within group if the data fed into it was grouped}
\usage{
Flag_From_Trend(DF, base_data, trend_data, sd_degree = 2.5)
}
\arguments{
\item{DF}{Dataframe containing selected columns}

\item{base_data}{Name of column containing Raw data}

\item{trend_data}{Name of generated trend data}

\item{sd_degree}{Threshold to flag deviant values as an outlier}
}
\value{
DF with a new column 'flagged_outlier' that contains if the column is an outlier
}
\description{
Flag values as outliers based on error from estimated trend
This function can be done within group if the data fed into it was grouped
}
