% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--data_prep--filter_tools--outlier_tools.R
\name{flagOutliers}
\alias{flagOutliers}
\title{Create column with Boolean based on a threashold}
\usage{
flagOutliers(DF, threshold, col = MessureRank, outputColName = FlaggedOutlier)
}
\arguments{
\item{DF}{Dataframe containing Column column
ranked_quantile_data @param threshold a numeric used to flag if its an outlier}

\item{col}{column being flagged based on threshold}

\item{outputColName}{name of flag column}
}
\value{
DF Dataframe with the extra column of if its flagged an outlier
}
\description{
Create column with Boolean based on a threashold
}
\examples{
data(example_data, package = "DSIWastewater")
example_data$site = "Madison"
df_data <- computeJumps(example_data)
ranked_data <- rankJumps(df_data)
 <- computeRankQuantiles(ranked_data)
flagOutliers(ranked_quantile_data, 9)
}
