% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--analysis_tools--time_offset--Offset_Analysis.R
\name{heatmapcorfunc}
\alias{heatmapcorfunc}
\title{Outputs a heatmap where the color is the r squared of wastewater data and center day + x many future days and y many past days
Helps inform Offset Analysis}
\usage{
heatmapcorfunc(cordata)
}
\arguments{
\item{cordata}{DF with geo_mean and conf_case columns}
}
\value{
ggplot plot object (heatmap)
}
\description{
Outputs a heatmap where the color is the r squared of wastewater data and center day + x many future days and y many past days
Helps inform Offset Analysis
}
