% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{HFGCase_data}
\alias{HFGCase_data}
\title{High frequency case data}
\format{
Rows: 788   Columns: 6
}
\usage{
data("HFGCase_data")
}
\description{
Data from the DHS of positive covid tests for HFG data
}
\keyword{datasets}
