% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{HFGWaste_data}
\alias{HFGWaste_data}
\title{High frequency Wastewater data set}
\format{
Rows: 3078  Columns: 20
}
\usage{
data("HFGWaste_data")
}
\description{
Data from the DHS of wastewater prevalence in selected Wisconsin cites where
high frequency testing was done
}
\details{
Raw file 'HFG_data_corrected_w_HF183&CrP_011422.xlsx'
From email Dagmara on Friday 2/25/2022
}
\keyword{datasets}
