% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--data_prep--smooth_tools--ts_trend_gen.R
\name{loessSmoothMod}
\alias{loessSmoothMod}
\title{loessSmoothMod
Add a column of the smoothed values using Loess}
\usage{
loessSmoothMod(
  DF,
  InVar = "N1",
  OutVar = "Loess",
  span = "guess",
  Filter = NULL
)
}
\arguments{
\item{DF}{DF we are adding the loess smooth col to}

\item{InVar}{The column to be smoothed}

\item{OutVar}{The name of the new column}

\item{span}{The span fed into loess smoothing. if it equals "guess" then it
if found using parameterGuess}

\item{Filter}{Prefilter using the value of a Filter col}
}
\value{
A DF with an extra col with a loesss smoothed version of InVar
}
\description{
loessSmoothMod
Add a column of the smoothed values using Loess
}
\examples{
data(Example_data, package = "DSIWastewater")
head(loessSmoothMod(Example_data))
}
