% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/library--analysis_tools--time_offset--Offset_Analysis.R
\name{OffsetHeatmap}
\alias{OffsetHeatmap}
\title{Outputs a heatmap of the offset for variant / time windows and population size / region
Must have pop_data, covarstarts, covarends loaded}
\usage{
OffsetHeatmap(method, timePeriods, wasterv, caserv, list, lod = FALSE, week)
}
\arguments{
\item{method}{Which analysis definds the offset (r squared, pearson, r squared offset, pearson offset, kendall offset, spearman offset)}

\item{timePeriods}{Size of time windows in months (if 0 uses variants)}

\item{wasterv}{DF of waste data must include: date, N1, N2}

\item{caserv}{DF of case data must include: date, conf_case}

\item{list}{y axis bins (population size or regions)}

\item{lod}{if true removes all values below LOD (default false)}

\item{week}{if true applies 7-day smoothing to case data}
}
\value{
ggplot plot object
}
\description{
Outputs a heatmap of the offset for variant / time windows and population size / region
Must have pop_data, covarstarts, covarends loaded
}
\examples{
data(WasteWater_data, package = "DSIWastewater")
data("Case_data", package = "DSIWastewater")
data(pop_data, package = "DSIWastewater")
OffsetHeatmap("kendall_offset",0,waste,cases,"pop",TRUE,"Both tests",TRUE)
OffsetHeatmap("kendall_offset",0,waste,cases,"pop",TRUE,"qPCR",TRUE)

}
