% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library--analysis_tools--lm_rand_forest--info.R
\name{OOB_MSE_num_trees}
\alias{OOB_MSE_num_trees}
\title{get OOB MSE vs number of forest in trees}
\usage{
OOB_MSE_num_trees(tree_model)
}
\arguments{
\item{tree_model}{random_linear_forest model to calculate the OOB mean squared error}
}
\value{
get dataframe of number of trees and OOB MSE
}
\description{
get OOB MSE vs number of forest in trees
}
\examples{
data("example_data", package = "DSIWastewater")
model <- random_linear_forest(example_data, 2, PMMoV ~ N1 + N2 | date + site)
OOB_MSE_num_trees(model)
}
