% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_rand_forest--info.R
\name{OOB_MSE_num_trees}
\alias{OOB_MSE_num_trees}
\title{get OOB MSE vs number of forest in trees}
\usage{
OOB_MSE_num_trees(tree_model)
}
\arguments{
\item{tree_model}{}
}
\value{
get dataframe of number of trees and OOB MSE
}
\description{
get OOB MSE vs number of forest in trees
}
\examples{
data("example_data", package = "DSIWastewater")
model <- random_linear_forest(example_data, 2, PMMoV ~ N1 + N2 | date + site)
gen_OOB_pred(model)
}
