% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{pop_data}
\alias{pop_data}
\title{City Population data}
\format{
Rows: 89   Columns: 2
}
\usage{
data("HFGCase_data")
}
\description{
dataframe containing info about each Wisconsin population info
}
\keyword{datasets}
