% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{Pop_data}
\alias{Pop_data}
\title{Sewer shed population data}
\format{
Rows: 89   Columns: 2
}
\usage{
data("Pop_data")
}
\description{
dataframe containing population information about each sewer shed in Wisconsin
}
\keyword{datasets}
