% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_tools--lm_rand_forest--use.R
\name{show,random_linear_forest-method}
\alias{show,random_linear_forest-method}
\title{display form for random_linear_forest class}
\usage{
\S4method{show}{random_linear_forest}(object)
}
\arguments{
\item{random_linear_forest}{object with class random_linear_forest}
}
\description{
display form for random_linear_forest class
}
\examples{
data("example_data", package = "DSIWastewater")
random_linear_forest(PMMoV ~ N1 + N2 | date + site)
}
