% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_algo.R
\name{uniqueVal}
\alias{uniqueVal}
\title{Find all unique values in the column selected}
\usage{
uniqueVal(Col, DF)
}
\arguments{
\item{Col}{Col being looked at}

\item{DF}{the DF containing Col}
}
\value{
a list of each unique col value
}
\description{
Find all unique values in the column selected
}
