% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{wastewater_data}
\alias{wastewater_data}
\title{Wastewater data set cleaned from the surveillance branch}
\format{
Rows: 8373 Columns: 30
}
\usage{
data("wastewater_data")
}
\description{
Data from the DHS of wastewater prevalence in Wisconsin cites
}
\details{
Raw file 'Shareable_Data_wastewater.csv'
From zip file attached to 4/21/2022 email from Matthew Schinwald

Processed into 'Shareable_Data_wastewater.edited.cleaned.csv'
In surveillance branch
}
\keyword{datasets}
