% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{WasteWater_data}
\alias{WasteWater_data}
\title{Wastewater data set}
\format{
Rows: 11,084 Columns: 62
}
\usage{
data("WasteWater_data")
}
\description{
Data from the DHS of wastewater prevalence in Wisconsin cites
}
\keyword{datasets}
